+++ CopilotKit/packages/react-ui/CHANGELOG.md
@@ -1,5 +1,35 @@
 # ui
 
+## 1.3.16-mme-sdk-js.2
+
+### Patch Changes
+
+- Ensure intermediate state config is sent as snake case
+- Updated dependencies
+  - @copilotkit/runtime-client-gql@1.3.16-mme-sdk-js.2
+  - @copilotkit/react-core@1.3.16-mme-sdk-js.2
+  - @copilotkit/shared@1.3.16-mme-sdk-js.2
+
+## 1.3.16-mme-sdk-js.1
+
+### Patch Changes
+
+- Update lockfile
+- Updated dependencies
+  - @copilotkit/runtime-client-gql@1.3.16-mme-sdk-js.1
+  - @copilotkit/react-core@1.3.16-mme-sdk-js.1
+  - @copilotkit/shared@1.3.16-mme-sdk-js.1
+
+## 1.3.16-mme-sdk-js.0
+
+### Patch Changes
+
+- Export LangGraph functions
+- Updated dependencies
+  - @copilotkit/react-core@1.3.16-mme-sdk-js.0
+  - @copilotkit/runtime-client-gql@1.3.16-mme-sdk-js.0
+  - @copilotkit/shared@1.3.16-mme-sdk-js.0
+
 ## 1.3.15
 
 ### Patch Changes